// frontend/src/lib/api/generated/model/githubComAbhishekPenDriveBackendInternalApiDtoDuplicateConflictPolicy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * pen-drive API
 * pen-drive backend API
 * OpenAPI spec version: 0.1.0
 */

export type GithubComAbhishekPenDriveBackendInternalApiDtoDuplicateConflictPolicy = typeof GithubComAbhishekPenDriveBackendInternalApiDtoDuplicateConflictPolicy[keyof typeof GithubComAbhishekPenDriveBackendInternalApiDtoDuplicateConflictPolicy];


export const GithubComAbhishekPenDriveBackendInternalApiDtoDuplicateConflictPolicy = {
  DuplicateConflictPolicyReject: 'reject',
  DuplicateConflictPolicyRename: 'rename',
  DuplicateConflictPolicyReplace: 'replace',
} as const;