// frontend/src/lib/api/generated/model/githubComAbhishekPenDriveBackendInternalApiDtoDuplicatePreviewItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * pen-drive API
 * pen-drive backend API
 * OpenAPI spec version: 0.1.0
 */

export interface GithubComAbhishekPenDriveBackendInternalApiDtoDuplicatePreviewItem {
  conflict?: boolean;
  existing_path?: string;
  rename_path?: string;
  requested_path?: string;
}