// frontend/src/lib/api/generated/model/githubComAbhishekPenDriveBackendInternalApiDtoDuplicatePreviewResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * pen-drive API
 * pen-drive backend API
 * OpenAPI spec version: 0.1.0
 */
import type { GithubComAbhishekPenDriveBackendInternalApiDtoDuplicatePreviewItem } from './githubComAbhishekPenDriveBackendInternalApiDtoDuplicatePreviewItem';

export interface GithubComAbhishekPenDriveBackendInternalApiDtoDuplicatePreviewResponse {
  has_conflicts?: boolean;
  impacted_paths?: string[];
  items?: GithubComAbhishekPenDriveBackendInternalApiDtoDuplicatePreviewItem[];
}