// frontend/src/lib/api/generated/model/githubComAbhishekPenDriveBackendInternalApiDtoMultipartUploadInitiateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * pen-drive API
 * pen-drive backend API
 * OpenAPI spec version: 0.1.0
 */

export interface GithubComAbhishekPenDriveBackendInternalApiDtoMultipartUploadInitiateResponse {
  key?: string;
  name?: string;
  part_size?: number;
  upload_id?: string;
}